Sales/requirements.txt
Flask
flask_sqlalchemy
flask_marshmallow
marshmallow
PyMySQL
SQLAlchemy
requests
PyJWT

